clean up and simplified twa logic
