+++ src/consts.ts
@@ -1,4 +1,5 @@
 import _ from "lodash";
+import twa from '@twa-dev/sdk'
 
 export const routes = {
   spaces: "/",
@@ -36,3 +37,5 @@ export const MOBILE_WIDTH = 768;
 
 export const ONE_WALLET_ONE_VOTE_URL =
   "https://github.com/orbs-network/ton-vote/blob/main/README.md#supported-strategies";
+
+export const isTwa = twa.initData
