+++ src/twa/index.ts
@@ -2,4 +2,3 @@ export * from './TwaConnectButton'
 export * from './TwaMainButtonWrapper'
 export * from './TwaConfirmVoteButton'
 export * from './TwaCastVoteButton'
-export * from './TwaBackButton'
